% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AurieLSHGaussian.R
\name{LSH_Gaussian}
\alias{LSH_Gaussian}
\title{Creates a Neighbourhood Using Locality Sensitive Hashing for Gaussian Projections}
\usage{
LSH_Gaussian(mydata, result9)
}
\arguments{
\item{mydata}{A data frame consisting of the data set without the class column}

\item{result9}{A column which consists of the class column}
}
\description{
This package uses Locality Sensitive Hashing and creates a Neighbourhood Graph for a datset and calculates the ARI value for the same. It uses Gaussian Random planes to decide the nature of a given point.
}
\examples{
LSH_Gaussian(iris[,-5],iris$Species)

}
